% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCH0.R
\name{MCH0}
\alias{MCH0}
\title{Monte Carlo estimation of the critial value of the S test}
\usage{
MCH0(
  X,
  A,
  b,
  tau = 0.5,
  alpha = 0.05,
  M = 1000,
  Rank = F,
  rescale = T,
  precalculated = T,
  homopowerrescaling = NA,
  q = 1
)
}
\arguments{
\item{X}{The covariate matrix.}

\item{A}{The linear constraint matrix. If \code{A="diff"} then the total variation test is performed.}

\item{b}{The constraint vector.}

\item{tau}{The quantile. The default value is 0.5.}

\item{alpha}{The test level. The default value is 0.05.}

\item{M}{The number of Monte Carlo runs. The default value is 1000.}

\item{Rank}{Logical value. If TRUE, the rankS test is performed; otherwise, the S test is performed. The default value is FALSE.}

\item{rescale}{Logical value. If TRUE, the matrix \code{A} and the vector \code{b} are rescaled according to the homopower rule; otherwise, \code{A} and \code{b} are not rescaled. The default value is TRUE.}

\item{precalculated}{Either set to TRUE (default value), in which case the precalculatedmatrices function is called,
otherwise, precalculated should be the output object of the precalculatedmatrices function.}

\item{homopowerrescaling}{Weights used to rescale the rows of \code{A} and \code{b}.}

\item{q}{The q/(q-1)-norm of the score used by the test statistic. Either 1 (default value corresponding to a LASSO penalty) or 2 (corresponding to a group LASSO penalty).}
}
\value{
\item{lambdas}{All M samples of the test statistic S under H0 simulated by Monte Carlo.}
\item{s_critval}{The estimated critical value \eqn{c_{\alpha}^{\tau}} of the test by taking the upper \eqn{(1-\alpha)}-quantile of the M samples of the test statistic S under H0.}
\item{homopowerrescaling}{Weights used to rescale the rows of \code{A} and \code{b}.}
}
\description{
Estimate the \eqn{c_{\alpha}^{\tau}} value by Monte Carlo by simulated the distribution of the test statistic S under H0.
}
\examples{
n=100
p=20
X=matrix(rnorm(n*p),n,p)
beta=rnorm(p)
A=diag(5)
A=cbind(A,matrix(0,5,p-5))
A[1,2]=A[2,3]=A[3,4]=A[4,5]=A[5,6]=-1
b=A\%*\%beta
MCH0(X,A,b)
}
