% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Stest.R
\name{plot.Stest}
\alias{plot.Stest}
\title{Plot for Stest}
\usage{
\method{plot}{Stest}(
  objStest,
  main = expression("Data"),
  xlab = expression("i"),
  ylab = expression(y[i]),
  ...
)
}
\arguments{
\item{objStest}{The object of Stest.}

\item{main}{The main for the data. The default value is "Data".}

\item{xlab}{The xlab for the data and estimated quantiles. The default value is "i".}

\item{ylab}{The ylab for the data. The default value is "\eqn{y_i}".}
}
\description{
Plot the density of lambdas simulated from Monte Carlon, quantile affine LASSO path. And if \code{A="diff"}, also the data, estimated jumps and estimated quantiles.
}
\examples{
n=100
p=20
X=matrix(rt(n*p, df=2),n,p)
beta=rnorm(p)
A=diag(5)
A=cbind(A,matrix(0,5,p-5))
A[1,2]=A[2,3]=A[3,4]=A[4,5]=A[5,6]=-1
b=A\%*\%beta
y=X\%*\%beta+rt(n,1)
obj=Stest(y,X,A,b)
plot.Stest(obj)
}
