% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precalculatedmatrices.R
\name{precalculatedmatrices}
\alias{precalculatedmatrices}
\title{Precalculated matrices to reduce the constrained regression}
\usage{
precalculatedmatrices(A, n = NA)
}
\arguments{
\item{A}{The linear constraint matrix. If \code{A="diff"} then the total variation test is performed.}

\item{n}{The sample size. The default value is NA. If \code{A="diff"}, then the sample size n must be provided.}
}
\value{
\item{A1}{The nonsingular matrix A1}
\item{A2}{The A2 matrix such that A=[A1, A2].}
\item{Xout_index}{The indexes of columns of A leading to a nonsingular matrix A1.}
\item{A1inv}{The inverse of A1.}
\item{AAtinv}{The inverse of \eqn{AA^{T}}.}
\item{AAtinvA}{\eqn{(AA^{T})^{-1}A}}
}
\description{
For the linear constraint \eqn{A\beta=b},
the matrix A must be decomposed into A=[A1, A2], where A1 is nonsingular.
Related matrices can also be precalculated to avoid repeating this operation several times.
This option is particularly useful for experienced users.
}
\examples{
p=20
A=diag(5)
A=cbind(A,matrix(0,5,p-5))
A[1,2]=A[2,3]=A[3,4]=A[4,5]=A[5,6]=-1
out=precalculatedmatrices(A)
##############
A="diff"
n=10
out=precalculatedmatrices(A,n)
}
