% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceXyfromAb.R
\name{reduceXyfromAb}
\alias{reduceXyfromAb}
\title{Transform a LAD optimization constrained to \eqn{A\beta=b} to an unconstrained LAD of lower dimension}
\usage{
reduceXyfromAb(y, X, A, b, precalculated = T)
}
\arguments{
\item{y}{The response vector.}

\item{X}{The covariate matrix.}

\item{A}{The linear constraint matrix. If \code{A="diff"} then the total variation test is performed.}

\item{b}{The constraint vector.}

\item{precalculated}{Either set to TRUE (default value), in which case the precalculatedmatrices function is called,
otherwise, precalculated should be the output object of the precalculatedmatrices function.}
}
\value{
\item{ynew}{The new y.}
\item{Xnew}{The new X.}
\item{A1}{The nonsingular matrix A1}
\item{A2}{The A2 matrix such that A=[A1, A2].}
\item{Xout_index}{The indexes of columns of A leading to a nonsingular matrix A1.}
\item{A1inv}{The inverse of A1.}
\item{AAtinv}{The inverse of \eqn{AA^{T}}.}
\item{AAtinvA}{\eqn{(AA^{T})^{-1}A}}
}
\description{
Returns the new y and new X via Gaussian elimitation of the equivalent LAD optimization without constraints.
}
\examples{
n=100
p=20
X=matrix(rnorm(n*p),n,p)
beta=rnorm(p)
A=diag(5)
A=cbind(A,matrix(0,5,p-5))
A[1,2]=A[2,3]=A[3,4]=A[4,5]=A[5,6]=-1
b=A\%*\%beta
y=X\%*\%beta+rnorm(n,1)
reduceXyfromAb(y,X,A,b)
}
