% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqaffineLASSO.R
\name{rqaffineLASSO}
\alias{rqaffineLASSO}
\title{Calcultion of regression quantile affine LASSO estimate.}
\usage{
rqaffineLASSO(y, X, A, b, tau = 0.5, lambda, q = 1)
}
\arguments{
\item{y}{The response vector.}

\item{X}{The covariate matrix.}

\item{A}{The linear constraint matrix. If \code{A="diff"} then the total variation test is performed.}

\item{b}{The constraint vector.}

\item{tau}{The quantile. The default value is 0.5.}

\item{lambda}{The penalty parameter.}

\item{q}{The q/(q-1)-norm of the score used by the test statistic. Either 1 (default value) or 2.}
}
\value{
\item{loss_value}{The loss value of regression quantile affine LASSO.}
\item{betahat}{The solution of regression quantile affine LASSO.}
}
\description{
For a given penalty parameter lambda, calculate the solution of the regression quantile affine LASSO optimization prob.
lem.
}
\examples{
n=100
p=20
X=matrix(rt(n*p, df=2),n,p)/sqrt(n)
beta=c(c(3,3,4,4),rnorm(p-4,sd=5))
A=diag(5)
A=cbind(A,matrix(0,5,p-5))
A[1,2]=A[2,3]=A[3,4]=A[4,5]=A[5,6]=-1
b=A\%*\%beta
y=X\%*\%beta+rt(n,1)
rqaffineLASSO(y,X,A,b,lambda=1)
}
