% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stest.R
\name{Stest}
\alias{Stest}
\title{S test}
\usage{
Stest(
  y,
  X,
  A,
  b,
  tau = 0.5,
  alpha = 0.05,
  M = 1000,
  Rank = F,
  rescale = T,
  outMCH0 = NA,
  precalculated = T,
  homopowerrescaling = NA,
  q = 1,
  nlambda = 20
)
}
\arguments{
\item{y}{The response vector.}

\item{X}{The covariate matrix.}

\item{A}{The linear constraint matrix. If \code{A="diff"} then the total variation test is performed.}

\item{b}{The constraint vector.}

\item{tau}{The quantile. The default value is 0.5.}

\item{alpha}{The test level. The default value is 0.05.}

\item{M}{The number of Monte Carlo runs. The default value is 1000.}

\item{Rank}{Logical value. If TRUE, the rankS test is performed; otherwise, the S test is performed. The default value is FALSE.}

\item{rescale}{Logical value. If TRUE, the matrix \code{A} and the vector \code{b} are rescaled according to the homopower rule; otherwise, \code{A} and \code{b} are not rescaled. The default value is TRUE.}

\item{outMCH0}{The Monte Carlo experiments outcome providing \eqn{c_{\alpha}^{\tau}} and all M sampled test statistics value. The default value is NA.}

\item{precalculated}{Either set to TRUE (default value), in which case the precalculatedmatrices function is called,
otherwise, precalculated should be the output object of the precalculatedmatrices function.}

\item{homopowerrescaling}{Weights used to rescale the rows of \code{A} and \code{b}.}

\item{q}{The q/(q-1)-norm of the score used by the test statistic. Either 1 (default value corresponding to a LASSO penalty) or 2 (corresponding to a group LASSO penalty).}

\item{nlambda}{The number of lambda values to compute the quantile affine LASSO path. The default is 20. For a reasonable plot, nlambda should be larger than 10 (forced minimal value). Set it to 0 if no need to plot the path.}
}
\value{
\item{test_value}{The test value of S test.}
\item{p_value}{The p value of S test.}
\item{lambda0data}{The lambda from the data.}
\item{critval}{The Monte Carlo experiments outcome of \eqn{c_{\alpha}^{\tau}}}
\item{tau}{The quantile.}
\item{alpha}{The test level.}
\item{q}{The q/(q-1)-norm of the score used by the test statistic.}
\item{betahat}{The betahat corresponding to the critial value.}
\item{homopowerrescaling}{Weights used to rescale the rows of \code{A} and \code{b}.}
\item{outMCH0}{\eqn{c_{\alpha}^{\tau}} and all M sampled test statistics value.}
\item{precalculated}{The object of the precalculatedmatrices function.}
\item{lambdas_seq}{Lambdas for the quantile affine LASSO path.}
\item{allbetahat}{All betahat corresponding to the lambdas_seq.}
\item{X}{The covariate matrix.}
\item{y}{The response vector.}
\item{A}{The linear constraint matrix.}
\item{b}{The constraint vector.}
\item{rescaledA}{The rescaled \code{A}.}
\item{rescaledb}{The rescaled \code{b}.}
}
\description{
Calculate the test function value (0 or 1) and p-value for the S test applied to the data for the desired linear null hypopthesis.
}
\examples{
n=100
p=20
X=matrix(rnorm(n*p),n,p)
beta=rnorm(p)
A=diag(5)
A=cbind(A,matrix(0,5,p-5))
A[1,2]=A[2,3]=A[3,4]=A[4,5]=A[5,6]=-1
b=A\%*\%beta
y=X\%*\%beta+rnorm(n,1)
Stest(y,X,A,b)
}
