% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.Stest.R
\name{summary.Stest}
\alias{summary.Stest}
\title{Summary for Stest}
\usage{
\method{summary}{Stest}(objStest)
}
\arguments{
\item{objStest}{The object of Stest.}
}
\description{
Explain some results of Stest object.
}
\examples{
n=100
p=20
X=matrix(rt(n*p, df=2),n,p)
beta=rnorm(p)
A=diag(5)
A=cbind(A,matrix(0,5,p-5))
A[1,2]=A[2,3]=A[3,4]=A[4,5]=A[5,6]=-1
b=A\%*\%beta
y=X\%*\%beta+rt(n,1)
obj=Stest(y,X,A,b)
summary.Stest(obj)
}
