% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStest.R
\name{summaryStest}
\alias{summaryStest}
\title{Summary for Stest}
\usage{
summaryStest(objStest)
}
\arguments{
\item{objStest}{The object of Stest.}
}
\description{
Reports the estimated p-value and the decision of the S Stest .
}
\examples{
n=100
p=20
X=matrix(rt(n*p, df=2),n,p)
beta=rnorm(p)
A=diag(5)
A=cbind(A,matrix(0,5,p-5))
A[1,2]=A[2,3]=A[3,4]=A[4,5]=A[5,6]=-1
b=A\%*\%beta
y=X\%*\%beta+rt(n,1)
out=Stest(y,X,A,b)
summaryStest(out)
}
