% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztfaql.R
\name{ztfaql}
\alias{ztfaql}
\title{Zero-thresholding function of quantile affine LASSO}
\usage{
ztfaql(
  y,
  X,
  A,
  b,
  tau = 0.5,
  Rank = F,
  outtype = "1",
  precalculated = T,
  q = 1
)
}
\arguments{
\item{y}{The response vector.}

\item{X}{The covariate matrix.}

\item{A}{The linear constraint matrix. If \code{A="diff"} then the total variation test is performed.}

\item{b}{The constraint vector.}

\item{tau}{The quantile. The default value is 0.5.}

\item{Rank}{Logical value. If TRUE, the rankS test is performed; otherwise, the S test is performed. The default value is FALSE.}

\item{outtype}{If set to 1, returns lambda0. If set to 2, returns lambda0 and indexmax.}

\item{precalculated}{Either set to TRUE (default value), in which case the precalculatedmatrices function is called,
otherwise, precalculated should be the output object of the precalculatedmatrices function.}

\item{q}{The q/(q-1)-norm of the score used by the test statistic. Either 1 (default value corresponding to a LASSO penalty) or 2 (corresponding to a group LASSO penalty).}
}
\value{
\item{lambda0}{The value of zero-thresholding function.}
\item{indexmax}{The index of max value.}
}
\description{
Get the smallest \eqn{\lambda} value for the quantile affine LASSO estimator to set its penalty to zero.
}
\examples{
n=100
p=20
X=matrix(rt(n*p, df=2),n,p)
beta=rnorm(p)
A=diag(5)
A=cbind(A,matrix(0,5,p-5))
A[1,2]=A[2,3]=A[3,4]=A[4,5]=A[5,6]=-1
b=A\%*\%beta
y=X\%*\%beta+rt(n,1)
ztf_aql(y,X,A,b)
}
