% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Stest.R
\name{plot.Stest}
\alias{plot.Stest}
\title{Plot for Stest}
\usage{
\method{plot}{Stest}(
  objStest,
  main = expression("Data"),
  xlab = expression("i"),
  ylab = expression(y[i]),
  ...
)
}
\arguments{
\item{objStest}{The object of Stest.}

\item{main}{The main for the data. The default value is "Data".}

\item{xlab}{The xlab for the data and estimated quantiles. The default value is "i".}

\item{ylab}{The ylab for the data. The default value is "\eqn{y_i}".}

\item{\dots}{Other graphical parameters to plot}
}
\description{
Plot the histogram density of lambdas simulated by Monte Carlo and the quantile affine LASSO path. And if \code{A="diff"}, additionally plots the time series data with estimated jump locations, as well as the estimated regression quantiles.
}
\examples{
n=100
p=20
X=matrix(rnorm(n*p),n,p)
beta=rnorm(p)
A=diag(5)
A=cbind(A,matrix(0,5,p-5))
A[1,2]=A[2,3]=A[3,4]=A[4,5]=A[5,6]=-1
b=A\%*\%beta
y=X\%*\%beta+rnorm(n,1)
out1=Stest(y,X,A,b,nlambda=0)
out2=Stest(y,X,A,b,nlambda=20)
plot(out1)
plot(out2)
}
